Introduce a deployable index
The desire of this course is manifold, the most prominent goal is to feed the passion and interest of new developers. The next most important goal is to build habits and skills that can positively benefit organizations (either OSS or an employer or both).
This index page is the first step in a move to make this repo deployable to a web host such as Vercel or Netlify. The ultimate plan is to make this repo able to be forked, deployed, and gradually updated as the learner progresses.